Cleaned up and reorganized acoustics code
